Created class Pair and completed card_id adjustment
Made the class 'Pair' as a subsituted for tuple(in a sense). Orignal meant to be a generic class but decied to make the class work for integer only. Pair is meant to be immutable, so
once it set it can't be change.

Card_ID make use of the Pair class to create a unique id
that consist of the card's Suit and card's Value enumeration
value's integer in the form of Suit's val, Value's val. As such
each card has a unique id that can't be duplicated as there can only be one of each card object totaling up to 52 unique card.